Make prevCapture persist into nested parses
This pulls the `prevCapture` out to be "global" state for the parses and persist into nested invocations of `nestedParse`. This allows rules to determine if they are truly at the beginning of a line (by seeing if `prevCapture` is either empty or ends with a newline). Before, the `prevCapture` would reset at each nesting, so it would only allow you to determine if you were at the beginning of the nested parse, but not necessarily the line.

There are other approaches to preserving this data (such as passing `prevCapture` into the recursive `nestedParse` call) but these all either created more functions at runtime or broke the API, requiring clients to change their method signatures.